i66: Extract custom Modal to its own component
